Exec: disable term resizes on windows
SIGWINCH does not exist, it breaks the build
