+++ pkg/koyeb/instances_exec_windows.go
@@ -1,0 +1,12 @@
+//go:build windows
+// +build windows
+
+package koyeb
+
+import (
+	"context"
+)
+
+func watchTermSize(ctx context.Context, s *StdStreams) <-chan *TerminalSize {
+	return nil
+}
